:recycle: refactor:Refatorar a estrutura atual da parte secretaria e aplicando componentização solicitada
