+++ civitas-frontend/src/app/dashboard/secretaria/_components/index.ts
@@ -1,0 +1,1 @@
+export { default as SecretariaPageContent } from "./SecretariaPageContent";
